Split out built-in cache service adapter defs into separate file
