+++ src/service.js
@@ -16,82 +16,21 @@ function httpEtagProvider () {
     'getItem',
     'removeItem',
     'removeAllItems'
+    // info method hard-coded
   ]
 
   var itemCacheMethods = [
     'set',
     'get',
     'remove'
+    // info method hard-coded
   ]
 
-  var cacheAdapters = {
-
-    'localStorage': {
-      createCache: angular.noop,
-      getCache: function () {
-        return localStorage
-      },
-      setItem: function (localStorage, itemKey, value) {
-        localStorage.setItem(itemKey, JSON.stringify(value))
-      },
-      getItem: function (localStorage, itemKey) {
-        return JSON.parse(localStorage.getItem(itemKey))
-      },
-      removeItem: function (localStorage, itemKey) {
-        localStorage.removeItem(itemKey)
-      },
-      removeAllItems: function (localStorage, itemKey) {
-        angular.forEach(localStorage, function (value, key) {
-          localStorage.removeItem(key)
-        })
-      }
-    },
-
-    'sessionStorage': {
-      createCache: angular.noop,
-      getCache: function () {
-        return sessionStorage
-      },
-      setItem: function (sessionStorage, itemKey, value) {
-        sessionStorage.setItem(itemKey, JSON.stringify(value))
-      },
-      getItem: function (sessionStorage, itemKey) {
-        return JSON.parse(sessionStorage.getItem(itemKey))
-      },
-      removeItem: function (sessionStorage, itemKey) {
-        sessionStorage.removeItem(itemKey)
-      },
-      removeAllItems: function (sessionStorage, itemKey) {
-        angular.forEach(sessionStorage, function (value, key) {
-          sessionStorage.removeItem(key)
-        })
-      }
-    },
-
-    '$cacheFactory': {
-      createCache: function ($cacheFactory, cacheId, options) {
-        $cacheFactory(cacheId, options)
-      },
-      getCache: function ($cacheFactory, cacheId) {
-        return $cacheFactory.get(cacheId)
-      },
-      setItem: function (cache, itemKey, value) {
-        cache.put(itemKey, value)
-      },
-      getItem: function (cache, itemKey) {
-        return cache.get(itemKey)
-      },
-      removeItem: function (cache, itemKey) {
-        cache.remove(itemKey)
-      },
-      removeAllItems: function (cache, itemKey) {
-        cache.removeAll()
-      }
-    }
-
-  }
+  // Built-in adapters defined in ./cacheServiceAdapters.js
+  var cacheAdapters = {}
   var cacheDefinitions = {}
 
+  // Cache config defaults
   var defaultCacheId = 'httpEtagCache'
   var defaultEtagCacheOptions = {
     deepCopy: false,
